Finalização exercício Api validações

+++ games/views.py
@@ -5,6 +5,7 @@ from rest_framework.response import Response
 from games.models import Game
 from games.serializers import GameSerializer
 
+
 @api_view(['GET', 'POST'])
 def game_list(request):
     if request.method == 'GET':
@@ -19,10 +20,10 @@ def game_list(request):
         return Response(game_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-@api_view(['GET', 'PUT', 'DELETE' ])
-def game_detail(request, pk):
+@api_view(['GET', 'PUT', 'DELETE'])
+def game_detail(request, id):
     try:
-        game = Game.objects.get(pk=pk)
+        game = Game.objects.get(id=id)
     except Game.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
     if request.method == 'GET':
@@ -35,7 +36,9 @@ def game_detail(request, pk):
             return Response(game_serializer.data)
         return Response(game_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     elif request.method == 'DELETE':
-        game.delete()
+        game_serializer = GameSerializer(game)
+        if game_serializer.validate_release_date(game.release_date):
+            game.delete()
     return Response(status=status.HTTP_204_NO_CONTENT)
 
 # @api_view(['GET', 'PUT', 'POST', 'DELETE' ])
